requirements.txt
wheel
azure-identity
